Refactor Codex into Client and Project classes, update tests
