+++ src/cleanlab_codex/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MIT
-from cleanlab_codex.codex import Codex
+from cleanlab_codex.client import Client
 from cleanlab_codex.codex_tool import CodexTool
 
-__all__ = ["Codex", "CodexTool"]
+__all__ = ["Client", "CodexTool"]
